requirements.txt
pymodbus
adafruit-circuitpython-ads1x15
PyYAML
pigpio
board
busio 